chore: Rewrite code according to audit (#394)
* deduplicate some constants

* renaming the buffers in the threshold comparison code

* code deduplication in chacha.cu

* remove clippy warnings for the case where the gpu_dependent flag is not set

* deduplicate chacha code

* rename kernels to more descriptive names

* rename the otp functions to chacha en-/decrypt

* remove the gpu_dependant flag again

* fix environment variable in test

* fix the overflow in chacha

* fix in deduplicated chacha gpu code

* replace int with size_t in kernel indices

* assert if not enough cuda threads are present

* Revert "assert if not enough cuda threads are present"

This reverts commit a2a5c60377d8ade92adb984f37b3a9cf600c5453.

* check if kernel can be launched

* deduplicate the check_max_grid_size function

* remove AES again and update some casts and constants

* remove allow(unused) in favor of feature flag

* fix the launch configs

* fix an error in new launch configs

* Unify number of default threads for the kernels

* remove clippy lint

+++ iris-mpc-gpu/src/dot/kernel.cu
@@ -4,9 +4,9 @@
 #define U8 unsigned char
 #define MAX_MATCHES_LEN 256
 
-extern "C" __global__ void xor_assign_u8(U8 *lhs, U8 *rhs, int n)
+extern "C" __global__ void xor_assign_u8(U8 *lhs, U8 *rhs, size_t n)
 {
-    int i = blockIdx.x * blockDim.x + threadIdx.x;
+    size_t i = blockIdx.x * blockDim.x + threadIdx.x;
     if (i < n)
     {
         lhs[i] ^= rhs[i];
